Crate siteverification1

Features

Handle the following Resources with ease from the central hub ...

web resource (delete, get, get token, insert, list, patch and update)

Everything else about the siteVerification API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.web_resource().patch(...).doit()
let r = hub.web_resource().get(...).doit()
let r = hub.web_resource().insert(...).doit()
let r = hub.web_resource().update(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
siteverification1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate siteverification1;
use siteverification1::SiteVerificationWebResourceResource;
use siteverification1::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use siteverification1::SiteVerification;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = SiteVerification::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: SiteVerificationWebResourceResource = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.web_resource().patch(&req, "id")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

SiteVerification
Central instance to access all SiteVerification related resource activities

SiteVerificationWebResourceGettokenRequest
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get token web resource (request)

SiteVerificationWebResourceGettokenRequestSite
The site for which a verification token will be generated.

This type is not used in any activity, and only used as part of another schema.

SiteVerificationWebResourceGettokenResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get token web resource (response)

SiteVerificationWebResourceListResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list web resource (response)

SiteVerificationWebResourceResource
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch web resource (request|response)

get web resource (response)

insert web resource (request|response)

update web resource (request|response)

SiteVerificationWebResourceResourceSite
The address and type of a site that is verified or will be verified.

This type is not used in any activity, and only used as part of another schema.

WebResourceDeleteMethodBuilder
Relinquish ownership of a website or domain.

WebResourceGetMethodBuilder
Get the most current data for a website or domain.

WebResourceGetTokenMethodBuilder
Get a verification token for placing on a website or domain.

WebResourceInsertMethodBuilder
Attempt verification of a website or domain.

WebResourceListMethodBuilder
Get the list of your verified websites and domains.

WebResourceMethodsBuilder
A builder providing access to all methods supported on webResource resources. It is not used directly, but through the SiteVerification hub.

WebResourcePatchMethodBuilder
Modify the list of owners for your website or domain. This method supports patch semantics.

WebResourceUpdateMethodBuilder
Modify the list of owners for your website or domain.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.